Move old css files along with their components
