+++ assets/src/components/books/detail/BookDetail.jsx
@@ -5,11 +5,12 @@ import Dialog from '@material-ui/core/Dialog';
 import Link from '@material-ui/core/Link';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogActions from '@material-ui/core/DialogActions';
-import '../../../../css/ModalBook.css';
 import Clear from '@material-ui/icons/Clear';
 import BookBorrowers from './BookBorrowers';
 import BookPublicationInfo from './BookPublicationInfo';
 import { BookPropType } from '../../../utils/propTypes';
+
+import './BookDetail.css';
 
 export default class BookDetail extends Component {
   constructor(props) {
